// src/app/core/interfaces-exam/exam-interfaces.ts
import { Metadata } from '../interfaces/common';

export interface Exam {
  _id: string;
  title: string;
  duration: number;
  subject: string;
  numberOfQuestions: number;
  active: boolean;
  createdAt: string;
}

export interface ExamsResponse {
  message: string;
  metadata: Metadata;
  exams: Exam[];
}

export interface ExamsAdaptResponse {
  message: string;
  exams: Exam[];
}